Refactor getting the contracts from safe-deployments
