+++ packages/safe-core-sdk/src/managers/contractManager.ts
@@ -1,32 +1,8 @@
-import {
-  EthAdapter,
-  GnosisSafeContract,
-  MultiSendContract,
-  SafeVersion
-} from '@gnosis.pm/safe-core-sdk-types'
+import { GnosisSafeContract, MultiSendContract } from '@gnosis.pm/safe-core-sdk-types'
 import { SAFE_LAST_VERSION } from '../contracts/config'
-import {
-  getMultiSendContractDeployment,
-  getSafeContractDeployment
-} from '../contracts/safeDeploymentContracts'
+import { getMultiSendContract, getSafeContract } from '../contracts/safeDeploymentContracts'
 import { SafeConfig } from '../Safe'
-import { ContractNetworkConfig, ContractNetworksConfig } from '../types'
-
-interface GetSafeContractInstanceProps {
-  ethAdapter: EthAdapter
-  safeVersion: SafeVersion
-  chainId: number
-  safeAddress: string
-  isL1SafeMasterCopy?: boolean
-  customContracts?: ContractNetworkConfig
-}
-
-interface GetMultiSendContractInstanceProps {
-  ethAdapter: EthAdapter
-  safeVersion: SafeVersion
-  chainId: number
-  customContracts?: ContractNetworkConfig
-}
+import { ContractNetworksConfig } from '../types'
 
 class ContractManager {
   #contractNetworks?: ContractNetworksConfig
@@ -56,24 +32,24 @@ class ContractManager {
     this.#contractNetworks = contractNetworks
     this.#isL1SafeMasterCopy = isL1SafeMasterCopy
 
-    const temporarySafeContract = await this.getSafeContract({
+    const temporarySafeContract = await getSafeContract({
       ethAdapter,
       safeVersion: SAFE_LAST_VERSION,
       chainId,
-      safeAddress,
       isL1SafeMasterCopy,
+      customSafeAddress: safeAddress,
       customContracts
     })
     const safeVersion = await temporarySafeContract.getVersion()
-    this.#safeContract = await this.getSafeContract({
+    this.#safeContract = await getSafeContract({
       ethAdapter,
       safeVersion,
       chainId,
-      safeAddress,
       isL1SafeMasterCopy,
+      customSafeAddress: safeAddress,
       customContracts
     })
-    this.#multiSendContract = await this.getMultiSendContract({
+    this.#multiSendContract = await getMultiSendContract({
       ethAdapter,
       safeVersion,
       chainId,
@@ -95,56 +71,6 @@ class ContractManager {
 
   get multiSendContract(): MultiSendContract {
     return this.#multiSendContract
-  }
-
-  private async getSafeContract({
-    ethAdapter,
-    safeVersion,
-    chainId,
-    safeAddress,
-    isL1SafeMasterCopy,
-    customContracts
-  }: GetSafeContractInstanceProps): Promise<GnosisSafeContract> {
-    const singletonDeployment = getSafeContractDeployment(
-      safeVersion,
-      chainId,
-      isL1SafeMasterCopy
-    )
-    const gnosisSafeContract = ethAdapter.getSafeContract({
-      safeVersion,
-      chainId,
-      singletonDeployment,
-      customContractAddress: safeAddress,
-      customContractAbi: customContracts?.safeMasterCopyAbi
-    })
-    const isContractDeployed = await ethAdapter.isContractDeployed(
-      gnosisSafeContract.getAddress()
-    )
-    if (!isContractDeployed) {
-      throw new Error('Safe Proxy contract is not deployed in the current network')
-    }
-    return gnosisSafeContract
-  }
-
-  private async getMultiSendContract({
-    ethAdapter,
-    safeVersion,
-    chainId,
-    customContracts
-  }: GetMultiSendContractInstanceProps): Promise<MultiSendContract> {
-    const multiSendDeployment = getMultiSendContractDeployment(safeVersion, chainId)
-    const multiSendContract = await ethAdapter.getMultiSendContract({
-      safeVersion,
-      chainId,
-      singletonDeployment: multiSendDeployment,
-      customContractAddress: customContracts?.multiSendAddress,
-      customContractAbi: customContracts?.multiSendAbi
-    })
-    const isContractDeployed = await ethAdapter.isContractDeployed(multiSendContract.getAddress())
-    if (!isContractDeployed) {
-      throw new Error('Multi Send contract is not deployed in the current network')
-    }
-    return multiSendContract
   }
 }
