+++ packages/safe-core-sdk/src/safeFactory/index.ts
@@ -6,12 +6,9 @@ import {
   TransactionOptions
 } from '@gnosis.pm/safe-core-sdk-types'
 import { SAFE_LAST_VERSION } from '../contracts/config'
-import {
-  getSafeContractDeployment,
-  getSafeProxyFactoryContractDeployment
-} from '../contracts/safeDeploymentContracts'
+import { getProxyFactoryContract, getSafeContract } from '../contracts/safeDeploymentContracts'
 import Safe from '../Safe'
-import { ContractNetworkConfig, ContractNetworksConfig } from '../types'
+import { ContractNetworksConfig } from '../types'
 import { EMPTY_DATA, ZERO_ADDRESS } from '../utils/constants'
 import { validateSafeAccountConfig } from './utils'
 
@@ -58,21 +55,6 @@ interface SafeFactoryInitConfig {
   contractNetworks?: ContractNetworksConfig
 }
 
-interface GetSafeContractInstanceProps {
-  ethAdapter: EthAdapter
-  safeVersion: SafeVersion
-  chainId: number
-  isL1SafeMasterCopy?: boolean
-  customContracts?: ContractNetworkConfig
-}
-
-interface GetProxyFactoryContractInstanceProps {
-  ethAdapter: EthAdapter
-  safeVersion: SafeVersion
-  chainId: number
-  customContracts?: ContractNetworkConfig
-}
-
 class SafeFactory {
   #contractNetworks?: ContractNetworksConfig
   #isL1SafeMasterCopy?: boolean
@@ -104,13 +86,13 @@ class SafeFactory {
     this.#contractNetworks = contractNetworks
     const chainId = await this.#ethAdapter.getChainId()
     const customContracts = contractNetworks?.[chainId]
-    this.#safeProxyFactoryContract = await this.getProxyFactoryContract({
+    this.#safeProxyFactoryContract = await getProxyFactoryContract({
       ethAdapter,
       safeVersion,
       chainId,
       customContracts
     })
-    this.#gnosisSafeContract = await this.getSafeContract({
+    this.#gnosisSafeContract = await getSafeContract({
       ethAdapter,
       safeVersion,
       chainId,
@@ -191,57 +173,6 @@ class SafeFactory {
       contractNetworks: this.#contractNetworks
     })
     return safe
-  }
-
-  private async getSafeContract({
-    ethAdapter,
-    safeVersion,
-    chainId,
-    isL1SafeMasterCopy,
-    customContracts
-  }: GetSafeContractInstanceProps): Promise<GnosisSafeContract> {
-    const singletonDeployment = getSafeContractDeployment(
-      safeVersion,
-      chainId,
-      isL1SafeMasterCopy
-    )
-    const gnosisSafeContract = ethAdapter.getSafeContract({
-      safeVersion,
-      chainId,
-      singletonDeployment,
-      customContractAddress: customContracts?.safeMasterCopyAddress,
-      customContractAbi: customContracts?.safeMasterCopyAbi
-    })
-    const isContractDeployed = await this.#ethAdapter.isContractDeployed(
-      gnosisSafeContract.getAddress()
-    )
-    if (!isContractDeployed) {
-      throw new Error('Safe Proxy contract is not deployed in the current network')
-    }
-    return gnosisSafeContract
-  }
-
-  private async getProxyFactoryContract({
-    ethAdapter,
-    safeVersion,
-    chainId,
-    customContracts
-  }: GetProxyFactoryContractInstanceProps): Promise<GnosisSafeProxyFactoryContract> {
-    const proxyFactoryDeployment = getSafeProxyFactoryContractDeployment(safeVersion, chainId)
-    const safeProxyFactoryContract = await ethAdapter.getSafeProxyFactoryContract({
-      safeVersion,
-      chainId,
-      singletonDeployment: proxyFactoryDeployment,
-      customContractAddress: customContracts?.safeProxyFactoryAddress,
-      customContractAbi: customContracts?.safeProxyFactoryAbi
-    })
-    const isContractDeployed = await this.#ethAdapter.isContractDeployed(
-      safeProxyFactoryContract.getAddress()
-    )
-    if (!isContractDeployed) {
-      throw new Error('Safe Proxy Factory contract is not deployed in the current network')
-    }
-    return safeProxyFactoryContract
   }
 }
